[feature] 5th iteration - added docstrings everywhere possible, separated out original code to folder & subfolders.
