+++ models/category.py
@@ -1,0 +1,6 @@
+
+# BONUS Category class
+class Category:
+    def __init__(self, category: str):
+        self.category = category
+        pass
